Move managers to their own folder
