+++ app/actions/remote/groups.ts
@@ -1,8 +1,9 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import NetworkManager from '@managers/network_manager';
+
 import {Client} from '@client/rest';
-import NetworkManager from '@init/network_manager';
 
 export const getGroupsForAutocomplete = async (serverUrl: string, channelId: string) => {
     let client: Client;
