+++ app/actions/remote/reactions.ts
@@ -1,11 +1,11 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import NetworkManager from '@managers/network_manager';
 import {Model} from '@nozbe/watermelondb';
 
 import {addRecentReaction} from '@actions/local/reactions';
 import DatabaseManager from '@database/manager';
-import NetworkManager from '@init/network_manager';
 import {getRecentPostsInChannel, getRecentPostsInThread} from '@queries/servers/post';
 import {queryReaction} from '@queries/servers/reaction';
 import {getCurrentChannelId, getCurrentUserId} from '@queries/servers/system';
