+++ app/actions/remote/systems.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import NetworkManager from '@managers/network_manager';
 import deepEqual from 'deep-equal';
 
 import {logError} from '@actions/remote/error';
@@ -8,7 +9,6 @@ import {forceLogoutIfNecessary} from '@actions/remote/session';
 import {SYSTEM_IDENTIFIERS} from '@constants/database';
 import DatabaseManager from '@database/manager';
 import {getServerCredentials} from '@init/credentials';
-import NetworkManager from '@init/network_manager';
 import {getCommonSystemValues} from '@queries/servers/system';
 
 import type ClientError from '@client/rest/error';
