+++ app/components/autocomplete/slash_suggestion/app_command_parser/app_command_parser.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import IntegrationsManager from '@managers/integrations_manager';
 import {Database} from '@nozbe/watermelondb';
 import {IntlShape} from 'react-intl';
 
@@ -9,7 +10,6 @@ import {fetchChannelById, fetchChannelByName, searchChannels} from '@actions/rem
 import {fetchUsersByIds, fetchUsersByUsernames, searchUsers} from '@actions/remote/user';
 import {AppCallResponseTypes, AppCallTypes, AppFieldTypes, COMMAND_SUGGESTION_ERROR} from '@constants/apps';
 import DatabaseManager from '@database/manager';
-import IntegrationsManager from '@init/integrations_manager';
 import {getChannelById, queryChannelsByNames} from '@queries/servers/channel';
 import {getCurrentTeamId} from '@queries/servers/system';
 import {getUserById, queryUsersByUsername} from '@queries/servers/user';
