+++ app/components/autocomplete/slash_suggestion/slash_suggestion.tsx
@@ -1,6 +1,8 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import analytics from '@managers/analytics';
+import IntegrationsManager from '@managers/integrations_manager';
 import {debounce} from 'lodash';
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {useIntl} from 'react-intl';
@@ -12,8 +14,6 @@ import {
 import {fetchSuggestions} from '@actions/remote/command';
 import {useServerUrl} from '@context/server';
 import {useTheme} from '@context/theme';
-import analytics from '@init/analytics';
-import IntegrationsManager from '@init/integrations_manager';
 import {makeStyleSheetFromTheme} from '@utils/theme';
 
 import {AppCommandParser} from './app_command_parser/app_command_parser';
