+++ app/managers/analytics.ts
