+++ app/managers/global_event_handler.ts
@@ -1,6 +1,9 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import * as analytics from '@managers/analytics';
+import NetworkManager from '@managers/network_manager';
+import WebsocketManager from '@managers/websocket_manager';
 import CookieManager, {Cookie} from '@react-native-cookies/cookies';
 import {Alert, DeviceEventEmitter, Linking, Platform} from 'react-native';
 import semver from 'semver';
@@ -10,12 +13,9 @@ import LocalConfig from '@assets/config.json';
 import {Events, Sso} from '@constants';
 import DatabaseManager from '@database/manager';
 import {DEFAULT_LOCALE, getTranslations, resetMomentLocale, t} from '@i18n';
-import * as analytics from '@init/analytics';
 import {getServerCredentials, removeServerCredentials} from '@init/credentials';
 import {getLaunchPropsFromDeepLink, relaunchApp} from '@init/launch';
-import NetworkManager from '@init/network_manager';
 import PushNotifications from '@init/push_notifications';
-import WebsocketManager from '@init/websocket_manager';
 import {getCurrentUser} from '@queries/servers/user';
 import EphemeralStore from '@store/ephemeral_store';
 import {LaunchType} from '@typings/launch';
