+++ app/managers/integrations_manager.ts
