+++ app/managers/network_manager.ts
