+++ app/screens/in_app_notification/icon.tsx
@@ -1,6 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import NetworkManager from '@managers/network_manager';
 import withObservables from '@nozbe/with-observables';
 import React from 'react';
 import {StyleSheet, View} from 'react-native';
@@ -9,7 +10,6 @@ import {of as of$} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
 
 import CompassIcon from '@components/compass_icon';
-import NetworkManager from '@init/network_manager';
 import {observeConfig} from '@queries/servers/system';
 import {observeUser} from '@queries/servers/user';
 import {WithDatabaseArgs} from '@typings/database/database';
