+++ app/screens/sso/sso_with_redirect_url.tsx
@@ -3,6 +3,7 @@
 
 import qs from 'querystring';
 
+import NetworkManager from '@managers/network_manager';
 import React, {useEffect, useState} from 'react';
 import {useIntl} from 'react-intl';
 import {Linking, Platform, Text, View} from 'react-native';
@@ -12,7 +13,6 @@ import urlParse from 'url-parse';
 
 import FormattedText from '@components/formatted_text';
 import {Sso} from '@constants';
-import NetworkManager from '@init/network_manager';
 import {buttonBackgroundStyle, buttonTextStyle} from '@utils/buttonStyles';
 import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
 import {typography} from '@utils/typography';
